refactor: Convert TripPlannerResultsSection to a function component
